+++ src/parser/tokenizers/transaction.rs
@@ -1,5 +1,5 @@
 use super::super::Rule;
-use crate::models::{Comment, PostingType, PriceType, Transaction, TransactionType};
+use crate::models::{Cleared, Comment, PostingType, PriceType, Transaction, TransactionType};
 use crate::parser::utils::{parse_date, parse_rational, parse_string};
 use crate::parser::Tokenizer;
 use chrono::NaiveDate;
@@ -31,8 +31,19 @@ impl<'a> Tokenizer<'a> {
                     transaction.effective_date =
                         Some(parse_date(part.into_inner().next().unwrap()));
                 }
-                Rule::status => {}
-                Rule::code => {}
+                Rule::status => {
+                    transaction.cleared = match part.as_str() {
+                        "!" => Cleared::NotCleared,
+                        "*" => Cleared::Cleared,
+                        x => panic!("Found '{}', expected '!' or '*'", x),
+                    }
+                }
+                Rule::code => {
+                    let mut code = part.as_str().chars();
+                    code.next();
+                    code.next_back();
+                    transaction.code = Some(code.as_str().trim().to_string())
+                }
                 Rule::description | Rule::automated_description => {
                     transaction.description = parse_string(part).trim().to_string();
                 }
@@ -219,12 +230,10 @@ fn parse_posting(
     }
     posting
 }
-
+#[cfg(test)]
 mod tests {
-
-    use crate::models::{TransactionStatus, Cleared};
-
     use super::*;
+    use crate::models::{Cleared, TransactionStatus};
     use crate::{parser::Tokenizer, CommonOpts};
 
     #[test]