+++ src/models.rs
@@ -395,6 +395,7 @@ impl ParsedLedger {
                 transaction.comments = parsed.comments.clone();
                 transaction.date = parsed.date;
                 transaction.effective_date = parsed.effective_date;
+                transaction.payee = parsed.payee.clone();
 
                 for comment in parsed.comments.iter() {
                     transaction.tags.append(&mut comment.get_tags());
@@ -520,4 +521,33 @@ pub trait HasAliases {
 
 pub trait FromDirective {
     fn is_from_directive(&self) -> bool;
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{parser::Tokenizer, CommonOpts};
+
+    #[test]
+    fn payee_with_pipe_issue_121() {
+        let mut tokenizer = Tokenizer::from(
+            "2022-05-13 ! (8760) Intereses | EstateGuru
+            EstateGuru               1.06 EUR
+            Ingresos:Rendimientos
+            "
+            .to_string(),
+        );
+        let options = CommonOpts::new();
+
+        let items = tokenizer.tokenize(&options);
+        dbg!(&items.transactions[0].payee);
+        let ledger = items.to_ledger(&options).unwrap();
+        let t = &ledger.transactions[0];
+        let payee = t.get_payee(&ledger.payees);
+        assert!(&ledger.payees.get("EstateGuru").is_ok());
+        dbg!(&payee);
+        dbg!(&t.payee);
+        dbg!(&ledger.payees);
+        assert!(payee.is_some());
+    }
 }