+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 Changelog file for dinero-rs project, a command line application for managing finances.
 ## [0.30.0] - xxx
-
+## Added
+- Show more info when loading the repl
 ## [0.29.1] - 2021-08-17
 ## Changed
 - small improvements on REPL interface