+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog file for dinero-rs project, a command line application for managing fi
 ## [0.30.0] - xxx
 ## Added
 - Show more info when loading the repl
+- Ability to [reload the journal](https://github.com/frosklis/dinero-rs/issues/116) 
 ## [0.29.1] - 2021-08-17
 ## Changed
 - small improvements on REPL interface