+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog file for dinero-rs project, a command line application for managing fi
 ## Added
 - Show more info when loading the repl
 - Ability to [reload the journal](https://github.com/frosklis/dinero-rs/issues/116) 
+## Fixed
+- Some payees were ```""``` (empty string) - not anymore :)
+
 ## [0.29.1] - 2021-08-17
 ## Changed
 - small improvements on REPL interface